+++ src/cards/Team/soccer-live-team.js
@@ -568,7 +568,7 @@ class SoccerLiveTeamCard extends LitElement {
           <div class="score-center">
             ${showScore
               ? html`<div class="score-numbers">${match.home_score} <span class="dash">-</span> ${match.away_score}</div>`
-              : html`<div class="score-vs">tegen</div>`
+              : html`<div class="score-vs">${this._t('match.vs')}</div>`
             }
             ${this._renderClock(match)}
           </div>